refactor Product ID type from Long to Integer in ProductController and ProductRepository

+++ src/ProductController.java
@@ -19,7 +19,7 @@ public class ProductController {
     }
 
     @PutMapping("/{id}")
-    public Product updateProductQuantity(@PathVariable Long id, @RequestBody Product updatedProduct) {
+    public Product updateProductQuantity(@PathVariable int id, @RequestBody Product updatedProduct) {
         return productRepository.findById(id).map(product -> {
             product.setQuantity(updatedProduct.getQuantity());
             return productRepository.save(product);
